build the edit issue page
