+++ app/issues/[id]/EditIssueButton.tsx
@@ -1,11 +1,12 @@
 import { Issue } from "@prisma/client";
 import { Pencil1Icon } from "@radix-ui/react-icons";
-import { Button } from "@react-email/components";
+import { Button } from "@radix-ui/themes";
 import Link from "next/link";
+import React from "react";
 
 export const EditIssueButton = ({ issue }: { issue: Issue }) => {
   return (
-    <Link href={`issues/${issue.id}/edit`}>
+    <Link href={`/issues/${issue.id}/edit`}>
       <Button>
         <Pencil1Icon />
         Edit Issue
