+++ app/issues/[id]/edit/page.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { IssueForm } from "../../_components/IssueForm";
+import prisma from "@/prisma/client";
+import { notFound } from "next/navigation";
+
+const EditIssue = async ({ params }: { params: { id: string } }) => {
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
+
+  if (!issue) notFound();
+
+  return <IssueForm issue={issue} />;
+};
+
+export default EditIssue;
